[quillmetadata] Move tests to /opt/. JB#57272
The /usr/lib part of tests.xml breaks on 64-bit devices.
Let's just move this to /opt which avoids ther problems and where
the majority of the tests are already.

Also removed the packagename replacement on the xml, the path is anyway
hard-coded to .cpp and we are not expected to build this with differet
package names.

Ut_regions not included in tests.xml but it doesn't pass now nor before.